model/requirements.txt
anyio==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
dataiku-api-client==9.0.1
fonttools==4.32.0
greenlet==1.1.2
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.19.0
threadpoolctl==3.1.0
toml==0.10.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
